[pkg/datadog] Refactor the API that provides metrics translator
