+++ pkg/datadog/config/metrics.go
@@ -7,12 +7,8 @@ import (
 	"encoding"
 	"fmt"
 
-	"github.com/DataDog/opentelemetry-mapping-go/pkg/otlp/attributes"
-	"github.com/DataDog/opentelemetry-mapping-go/pkg/otlp/attributes/source"
 	otlpmetrics "github.com/DataDog/opentelemetry-mapping-go/pkg/otlp/metrics"
-	"go.opentelemetry.io/collector/component"
 	"go.opentelemetry.io/collector/config/confignet"
-	"go.opentelemetry.io/collector/featuregate"
 )
 
 // MetricsConfig defines the metrics exporter specific configuration options
@@ -212,29 +208,10 @@ type MetricsExporterConfig struct {
 	InstrumentationScopeMetadataAsTags bool `mapstructure:"instrumentation_scope_metadata_as_tags"`
 }
 
-var metricRemappingDisableddFeatureGate = featuregate.GlobalRegistry().MustRegister(
-	"exporter.datadogexporter.metricremappingdisabled",
-	featuregate.StageAlpha,
-	featuregate.WithRegisterDescription("When enabled the Datadog Exporter remaps OpenTelemetry semantic conventions to Datadog semantic conventions. This feature gate is only for internal use."),
-	featuregate.WithRegisterReferenceURL("https://docs.datadoghq.com/opentelemetry/schema_semantics/metrics_mapping/"),
-)
-
-// isMetricRemappingDisabled returns true if the datadogexporter should generate Datadog-compliant metrics from OpenTelemetry metrics
-func isMetricRemappingDisabled() bool {
-	return metricRemappingDisableddFeatureGate.IsEnabled()
-}
-
-// TranslatorFromConfig creates a new metrics translator from the exporter
-func TranslatorFromConfig(set component.TelemetrySettings, mcfg MetricsConfig, attrsTranslator *attributes.Translator, sourceProvider source.Provider, statsOut chan []byte) (*otlpmetrics.Translator, error) {
+// ToTranslatorOpts returns a list of metrics translator options from the metrics config
+func (mcfg MetricsConfig) ToTranslatorOpts() []otlpmetrics.TranslatorOption {
 	options := []otlpmetrics.TranslatorOption{
 		otlpmetrics.WithDeltaTTL(mcfg.DeltaTTL),
-		otlpmetrics.WithFallbackSourceProvider(sourceProvider),
-	}
-
-	if isMetricRemappingDisabled() {
-		set.Logger.Warn("Metric remapping is disabled in the Datadog exporter. OpenTelemetry metrics must be mapped to Datadog semantics before metrics are exported to Datadog (ex: via a processor).")
-	} else {
-		options = append(options, otlpmetrics.WithRemapping())
 	}
 
 	if mcfg.HistConfig.SendAggregations {
@@ -262,6 +239,5 @@ func TranslatorFromConfig(set component.TelemetrySettings, mcfg MetricsConfig, a
 	options = append(options, otlpmetrics.WithInitialCumulMonoValueMode(
 		otlpmetrics.InitialCumulMonoValueMode(mcfg.SumConfig.InitialCumulativeMonotonicMode)))
 
-	options = append(options, otlpmetrics.WithStatsOut(statsOut))
-	return otlpmetrics.NewTranslator(set, attrsTranslator, options...)
+	return options
 }
